refactor: separate exports for node and web db connectors
