+++ packages/database/package.json
@@ -5,7 +5,8 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
-    "dist"
+    "dist",
+    "dist/web.js"
   ],
   "_moduleAliases": {
     "~database": "./src"
