+++ packages/database/src/index.ts
@@ -1,17 +1,8 @@
 import Web3 from 'web3'
-import schema from './schema'
 
+import schema from './schema'
 import { DB } from './types'
 import { Config } from './schema.types'
-
-export { SQLiteConnector } from './connectors/sqlite'
-export { SQLiteMemoryConnector } from './connectors/sqlite-memory'
-export { PostgresConnector } from './connectors/postgres'
-export { IndexedDBConnector } from './connectors/indexed-db'
-export { DB, TableData } from './types'
-export * from './schema.types'
-export * from './preset'
-export { schema }
 
 interface L1Contract {
   getConfig(): Promise<Config>
@@ -46,3 +37,8 @@ export enum TreeSpecies {
 }
 
 export const NULLIFIER_TREE_ID = 'nullifier-tree'
+
+export { DB, TableData } from './types'
+export * from './schema.types'
+export * from './preset'
+export { schema }
