+++ packages/database/src/web.ts
@@ -1,0 +1,2 @@
+export { IndexedDBConnector } from './connectors/indexed-db'
+export * from '.'
