+++ packages/tree/src/utxo-tree.ts
@@ -3,12 +3,8 @@ import {
   DB,
   LightTree,
   TreeSpecies,
-  SQLiteMemoryConnector,
-  schema,
 } from '@zkopru/database'
 import { ZkAddress } from '@zkopru/transaction'
-import { v4 } from 'uuid'
-import { genesisRoot, poseidonHasher } from './hasher'
 import {
   LightRollUpTree,
   TreeMetadata,
@@ -87,38 +83,5 @@ export class UtxoTree extends LightRollUpTree<Fp> {
       },
       config,
     })
-  }
-
-  static async sample(depth: number): Promise<{ tree: UtxoTree; db: DB }> {
-    const utxoTreeMetadata = {
-      id: v4(),
-      index: 1,
-      species: TreeSpecies.UTXO,
-      start: Fp.from(0),
-      end: Fp.from(0),
-    }
-    const utxoTreeConfig: TreeConfig<Fp> = {
-      hasher: poseidonHasher(depth),
-      forceUpdate: true,
-      fullSync: true,
-    }
-    const preHashes = poseidonHasher(depth).preHash
-    const utxoTreeInitialData = {
-      root: genesisRoot(poseidonHasher(depth)),
-      index: Fp.zero,
-      siblings: preHashes.slice(0, -1),
-    }
-    const mockupDB = await SQLiteMemoryConnector.create(schema)
-    const utxoTree = new UtxoTree({
-      db: mockupDB,
-      metadata: utxoTreeMetadata,
-      data: utxoTreeInitialData,
-      config: utxoTreeConfig,
-    })
-    await utxoTree.init()
-    return {
-      tree: utxoTree,
-      db: mockupDB,
-    }
   }
 }
